refactor: move create new task
