+++ batcher/aligned-batcher/src/eth/mod.rs
@@ -1,14 +1,12 @@
-use std::iter::repeat;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use aligned_sdk::eth::batcher_payment_service::{BatcherPaymentServiceContract, SignatureData};
+use aligned_sdk::eth::batcher_payment_service::BatcherPaymentServiceContract;
 use ethers::prelude::k256::ecdsa::SigningKey;
 use ethers::prelude::*;
 use gas_escalator::{Frequency, GeometricGasPrice};
-use log::info;
 
-use crate::{config::ECDSAConfig, types::errors::BatcherError};
+use crate::config::ECDSAConfig;
 
 #[derive(Debug, Clone, EthEvent)]
 pub struct BatchVerified {
@@ -38,42 +36,6 @@ pub fn get_provider(eth_rpc_url: String) -> Result<Provider<RetryClient<Http>>, 
     Ok(Provider::<RetryClient<Http>>::new(client))
 }
 
-pub async fn create_new_task(
-    payment_service: &BatcherPaymentService,
-    batch_merkle_root: [u8; 32],
-    batch_data_pointer: String,
-    leaves: Vec<[u8; 32]>,
-    signatures: Vec<SignatureData>,
-    gas_for_aggregator: U256,
-    gas_per_proof: U256,
-) -> Result<TransactionReceipt, BatcherError> {
-    // pad leaves to next power of 2
-    let padded_leaves = pad_leaves(leaves);
-
-    let call = payment_service.create_new_task(
-        batch_merkle_root,
-        batch_data_pointer,
-        padded_leaves,
-        signatures,
-        gas_for_aggregator,
-        gas_per_proof,
-    );
-
-    info!("Creating task for: {:x?}", batch_merkle_root);
-
-    let pending_tx = call
-        .send()
-        .await
-        .map_err(|e| BatcherError::TaskCreationError(e.to_string()))?;
-
-    let receipt = pending_tx
-        .await
-        .map_err(|_| BatcherError::TransactionSendError)?
-        .ok_or(BatcherError::ReceiptNotFoundError)?;
-
-    Ok(receipt)
-}
-
 pub async fn get_batcher_payment_service(
     provider: Provider<RetryClient<Http>>,
     ecdsa_config: ECDSAConfig,
@@ -98,13 +60,4 @@ pub async fn get_batcher_payment_service(
         BatcherPaymentService::new(H160::from_str(contract_address.as_str())?, signer);
 
     Ok(service_manager)
-}
-
-fn pad_leaves(leaves: Vec<[u8; 32]>) -> Vec<[u8; 32]> {
-    let leaves_len = leaves.len();
-    let last_leaf = leaves[leaves_len - 1];
-    leaves
-        .into_iter()
-        .chain(repeat(last_leaf).take(leaves_len.next_power_of_two() - leaves_len))
-        .collect()
 }
